+++ convex/entities.ts
@@ -32,6 +32,21 @@ import {
   verifyEntityInUnsavedRuleSet,
 } from "./copyOnWrite";
 import {
+  appointmentTypeResultValidator,
+  baseScheduleBatchResultValidator,
+  baseScheduleCreatePayloadValidator,
+  baseSchedulePayloadValidator,
+  baseScheduleResultValidator,
+  deletePractitionerWithDependenciesResultValidator,
+  expectedDraftRevisionValidator,
+  locationResultValidator,
+  practitionerDependencySnapshotValidator,
+  practitionerResultValidator,
+  replaceBaseScheduleSetResultValidator,
+  restorePractitionerWithDependenciesResultValidator,
+  ruleResultValidator,
+} from "./entities.validators";
+import {
   type FollowUpPlan,
   followUpStepValidator,
   validateFollowUpPlan,
@@ -56,166 +71,6 @@ type DatabaseWriter = GenericDatabaseWriter<DataModel>;
 // ================================
 // SHARED TYPES
 // ================================
-
-const appointmentTypeResultValidator = v.object({
-  draftRevision: v.number(),
-  entityId: v.id("appointmentTypes"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const practitionerResultValidator = v.object({
-  draftRevision: v.number(),
-  entityId: v.id("practitioners"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const locationResultValidator = v.object({
-  draftRevision: v.number(),
-  entityId: v.id("locations"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const baseScheduleResultValidator = v.object({
-  draftRevision: v.number(),
-  entityId: v.id("baseSchedules"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const baseScheduleBatchResultValidator = v.object({
-  createdScheduleIds: v.array(v.id("baseSchedules")),
-  draftRevision: v.number(),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const ruleResultValidator = v.object({
-  draftRevision: v.number(),
-  entityId: v.id("ruleConditions"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const baseSchedulePayloadValidator = v.object({
-  breakTimes: v.optional(
-    v.array(
-      v.object({
-        end: v.string(),
-        start: v.string(),
-      }),
-    ),
-  ),
-  dayOfWeek: v.number(),
-  endTime: v.string(),
-  lineageKey: v.id("baseSchedules"),
-  locationId: v.id("locations"),
-  practitionerId: v.id("practitioners"),
-  startTime: v.string(),
-});
-
-const baseScheduleCreatePayloadValidator = v.object({
-  breakTimes: v.optional(
-    v.array(
-      v.object({
-        end: v.string(),
-        start: v.string(),
-      }),
-    ),
-  ),
-  dayOfWeek: v.number(),
-  endTime: v.string(),
-  lineageKey: v.optional(v.id("baseSchedules")),
-  locationId: v.id("locations"),
-  practitionerId: v.id("practitioners"),
-  startTime: v.string(),
-});
-
-const replaceBaseScheduleSetResultValidator = v.object({
-  appliedSchedules: v.array(
-    v.object({
-      breakTimes: v.optional(
-        v.array(
-          v.object({
-            end: v.string(),
-            start: v.string(),
-          }),
-        ),
-      ),
-      dayOfWeek: v.number(),
-      endTime: v.string(),
-      entityId: v.id("baseSchedules"),
-      lineageKey: v.id("baseSchedules"),
-      locationId: v.id("locations"),
-      locationLineageKey: v.id("locations"),
-      practitionerId: v.id("practitioners"),
-      practitionerLineageKey: v.id("practitioners"),
-      startTime: v.string(),
-    }),
-  ),
-  createdScheduleIds: v.array(v.id("baseSchedules")),
-  deletedScheduleIds: v.array(v.id("baseSchedules")),
-  draftRevision: v.number(),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const practitionerBaseScheduleSnapshotValidator = v.object({
-  breakTimes: v.optional(
-    v.array(
-      v.object({
-        end: v.string(),
-        start: v.string(),
-      }),
-    ),
-  ),
-  dayOfWeek: v.number(),
-  endTime: v.string(),
-  lineageKey: v.id("baseSchedules"),
-  locationId: v.id("locations"),
-  locationLineageKey: v.id("locations"),
-  locationOriginId: v.optional(v.id("locations")),
-  startTime: v.string(),
-});
-
-const practitionerSnapshotValidator = v.object({
-  id: v.id("practitioners"),
-  lineageKey: v.id("practitioners"),
-  name: v.string(),
-  tags: v.optional(v.array(v.string())),
-});
-
-const practitionerAppointmentTypePatchValidator = v.object({
-  action: v.union(v.literal("delete"), v.literal("patch")),
-  afterAllowedPractitionerIds: v.array(v.id("practitioners")),
-  appointmentTypeId: v.id("appointmentTypes"),
-  beforeAllowedPractitionerIds: v.array(v.id("practitioners")),
-  duration: v.optional(v.number()),
-  lineageKey: v.id("appointmentTypes"),
-  name: v.optional(v.string()),
-});
-
-const practitionerConditionPatchValidator = v.object({
-  afterValueIds: v.array(v.string()),
-  beforeValueIds: v.array(v.string()),
-  conditionId: v.id("ruleConditions"),
-});
-
-const practitionerDependencySnapshotValidator = v.object({
-  appointmentTypePatches: v.array(practitionerAppointmentTypePatchValidator),
-  baseSchedules: v.array(practitionerBaseScheduleSnapshotValidator),
-  practitioner: practitionerSnapshotValidator,
-  practitionerConditionPatches: v.array(practitionerConditionPatchValidator),
-});
-
-const deletePractitionerWithDependenciesResultValidator = v.object({
-  draftRevision: v.number(),
-  ruleSetId: v.id("ruleSets"),
-  snapshot: practitionerDependencySnapshotValidator,
-});
-
-const restorePractitionerWithDependenciesResultValidator = v.object({
-  draftRevision: v.number(),
-  restoredPractitionerId: v.id("practitioners"),
-  ruleSetId: v.id("ruleSets"),
-});
-
-const expectedDraftRevisionValidator = v.union(v.number(), v.null());
 
 async function finalizeDraftMutation(
   db: DatabaseWriter,
