+++ src/components/calendar/use-calendar-logic-helpers.ts
@@ -1,0 +1,50 @@
+import type { RefObject } from "react";
+
+import { err, ok, type Result } from "neverthrow";
+import { Temporal } from "temporal-polyfill";
+
+import type { Id } from "../../../convex/_generated/dataModel";
+
+import { invalidStateError } from "../../utils/frontend-errors";
+
+export const TIMEZONE = "Europe/Berlin";
+
+export interface BlockedSlotConversionOptions {
+  endISO?: string;
+  locationId?: Id<"locations">;
+  practitionerId?: Id<"practitioners">;
+  startISO?: string;
+  title?: string;
+}
+
+export interface SimulationConversionOptions {
+  columnOverride?: string;
+  durationMinutes?: number;
+  endISO?: string;
+  locationId?: Id<"locations">;
+  practitionerId?: Id<"practitioners">;
+  startISO?: string;
+}
+
+export function handleEditBlockedSlot(
+  blockedSlotId: string,
+  justFinishedResizingRef: RefObject<null | string>,
+): boolean {
+  if (justFinishedResizingRef.current === blockedSlotId) {
+    return false;
+  }
+  return true;
+}
+
+export function parsePlainTimeResult(
+  value: string,
+  source: string,
+): Result<Temporal.PlainTime, ReturnType<typeof invalidStateError>> {
+  try {
+    return ok(Temporal.PlainTime.from(value));
+  } catch (error) {
+    return err(
+      invalidStateError(`Invalid time format: ${value}`, source, error),
+    );
+  }
+}
